[web-server] Custom Http Webserver
Created a basic Http server with the following components (mainly structs) -
a. Server : The http server listening on a specified port and servering client requests concurrently.
b. Request : Captures the Http request (server side)
c. RequestParser : Parses the byte slice into a Request
d. Response : The http response sent to the client
e. ResponseWriter : Serializes and writes the Response to the socket.
f. Registry : Registers the routes and handlers. (a very basic implementation)

The file names are 1:1 mapping for the above resources.
NOTE : "common.go" contains functions and Handlers which are common/constant.